+++ typings/jsforce/jsforce.d.ts
@@ -135,7 +135,7 @@ declare module 'jsforce/index' {
             upsert(foo: string, bar: any): Promise<any>;
         }
         interface Bulk {
-            load(type: string, operation: string, records: Array<{}>): Promise<any>;
+            load(type: string, operation: string, csvStream: any, callback: (err, rets) => any);
         }
         // Connection
         interface Connection {